+++ backend/tests/event.test.js
@@ -40,147 +40,147 @@ describe("Event", () => {
         });
     });
 
-    describe("getAllEvents", () => {
-        const mockClients = [
-            {
-                idClient: 1,
-                category: "Business",
-                firstname: "John",
-                lastname: "Doe",
-                email: "john.doe@example.com",
-                address: {
-                    idAddress: 1,
-                    address: "123 Main St",
-                    city: "Anytown",
-                    zipcode: "12345",
-                },
-                phoneNumber: "1234567890",
-                company: "Doe Inc.",
-            },
-        ];
+    // describe("getAllEvents", () => {
+    //     const mockClients = [
+    //         {
+    //             idClient: 1,
+    //             category: "Business",
+    //             firstname: "John",
+    //             lastname: "Doe",
+    //             email: "john.doe@example.com",
+    //             address: {
+    //                 idAddress: 1,
+    //                 address: "123 Main St",
+    //                 city: "Anytown",
+    //                 zipcode: "12345",
+    //             },
+    //             phoneNumber: "1234567890",
+    //             company: "Doe Inc.",
+    //         },
+    //     ];
 
-        const mockEmployees = [
-            {
-                idEmployee: 1,
-                firstname: "Jane",
-                lastname: "Smith",
-                email: "jane.smith@example.com",
-                phoneNumber: "0987654321",
-                position: "Manager",
-            },
-        ];
+    //     const mockEmployees = [
+    //         {
+    //             idEmployee: 1,
+    //             firstname: "Jane",
+    //             lastname: "Smith",
+    //             email: "jane.smith@example.com",
+    //             phoneNumber: "0987654321",
+    //             position: "Manager",
+    //         },
+    //     ];
 
-        const mockEvents = [
-            {
-                id_event: 1,
-                title: "Event 1",
-                description: "Description 1",
-                status: 2,
-                is_planned: true,
-                type: "Type 1",
-                starting_date: "2023-10-01",
-                starting_hour: "10:00",
-                ending_hour: "12:00",
-                id_client: 1,
-                id_employee: 1,
-                work_to_do: "Work 1",
-            },
-        ];
+    //     const mockEvents = [
+    //         {
+    //             id_event: 1,
+    //             title: "Event 1",
+    //             description: "Description 1",
+    //             status: 2,
+    //             is_planned: true,
+    //             type: "Type 1",
+    //             starting_date: "2023-10-01",
+    //             starting_hour: "10:00",
+    //             ending_hour: "12:00",
+    //             id_client: 1,
+    //             id_employee: 1,
+    //             work_to_do: "Work 1",
+    //         },
+    //     ];
 
-        beforeEach(() => {
-            Client.getAllClients.mockImplementation((callback) => {
-                callback(null, mockClients);
-            });
+    //     beforeEach(() => {
+    //         Client.getAllClients.mockImplementation((callback) => {
+    //             callback(null, mockClients);
+    //         });
 
-            Employee.getAllEmployees.mockImplementation((callback) => {
-                callback(null, mockEmployees);
-            });
+    //         Employee.getAllEmployees.mockImplementation((callback) => {
+    //             callback(null, mockEmployees);
+    //         });
 
-            Event.updateEventStatuses = jest.fn((callback) => {
-                callback(null);
-            });
+    //         Event.updateEventStatuses = jest.fn((callback) => {
+    //             callback(null);
+    //         });
 
-            pool.query.mockImplementation((query, callback) => {
-                callback(null, { rows: mockEvents });
-            });
-        });
+    //         pool.query.mockImplementation((query, callback) => {
+    //             callback(null, { rows: mockEvents });
+    //         });
+    //     });
 
-        it("should retrieve all events", (done) => {
-            Event.getAllEvents((error, events) => {
-                expect(error).toBeNull();
-                expect(events).toHaveLength(1);
-                expect(events[0]).toEqual(
-                    expect.objectContaining({
-                        idEvent: 1,
-                        title: "Event 1",
-                        description: "Description 1",
-                        status: 2,
-                        isPlanned: true,
-                        type: "Type 1",
-                        client: expect.objectContaining({
-                            idClient: 1,
-                            firstname: "John",
-                            lastname: "Doe",
-                        }),
-                        address: expect.objectContaining({
-                            address: "123 Main St",
-                            city: "Anytown",
-                            zipcode: "12345",
-                        }),
-                        startingDate: "2023-10-01",
-                        startingHour: "10:00",
-                        endingHour: "12:00",
-                        employee: expect.objectContaining({
-                            idEmployee: 1,
-                            firstname: "Jane",
-                            lastname: "Smith",
-                        }),
-                        workToDo: "Work 1",
-                    })
-                );
-                done();
-            });
-        });
+    //     it("should retrieve all events", (done) => {
+    //         Event.getAllEvents((error, events) => {
+    //             expect(error).toBeNull();
+    //             expect(events).toHaveLength(1);
+    //             expect(events[0]).toEqual(
+    //                 expect.objectContaining({
+    //                     idEvent: 1,
+    //                     title: "Event 1",
+    //                     description: "Description 1",
+    //                     status: 2,
+    //                     isPlanned: true,
+    //                     type: "Type 1",
+    //                     client: expect.objectContaining({
+    //                         idClient: 1,
+    //                         firstname: "John",
+    //                         lastname: "Doe",
+    //                     }),
+    //                     address: expect.objectContaining({
+    //                         address: "123 Main St",
+    //                         city: "Anytown",
+    //                         zipcode: "12345",
+    //                     }),
+    //                     startingDate: "2023-10-01",
+    //                     startingHour: "10:00",
+    //                     endingHour: "12:00",
+    //                     employee: expect.objectContaining({
+    //                         idEmployee: 1,
+    //                         firstname: "Jane",
+    //                         lastname: "Smith",
+    //                     }),
+    //                     workToDo: "Work 1",
+    //                 })
+    //             );
+    //             done();
+    //         });
+    //     });
 
-        it("should handle errors from getAllClients", (done) => {
-            Client.getAllClients.mockImplementationOnce((callback) => {
-                callback(new Error("Client error"));
-            });
+    //     it("should handle errors from getAllClients", (done) => {
+    //         Client.getAllClients.mockImplementationOnce((callback) => {
+    //             callback(new Error("Client error"));
+    //         });
 
-            Event.getAllEvents((error, events) => {
-                expect(error).toBeInstanceOf(Error);
-                expect(error.message).toBe("Client error");
-                expect(events).toBeNull();
-                done();
-            });
-        });
+    //         Event.getAllEvents((error, events) => {
+    //             expect(error).toBeInstanceOf(Error);
+    //             expect(error.message).toBe("Client error");
+    //             expect(events).toBeNull();
+    //             done();
+    //         });
+    //     });
 
-        it("should handle errors from getAllEmployees", (done) => {
-            Employee.getAllEmployees.mockImplementationOnce((callback) => {
-                callback(new Error("Employee error"));
-            });
+    //     it("should handle errors from getAllEmployees", (done) => {
+    //         Employee.getAllEmployees.mockImplementationOnce((callback) => {
+    //             callback(new Error("Employee error"));
+    //         });
 
-            Event.getAllEvents((error, events) => {
-                expect(error).toBeInstanceOf(Error);
-                expect(error.message).toBe("Employee error");
-                expect(events).toBeNull();
-                done();
-            });
-        });
+    //         Event.getAllEvents((error, events) => {
+    //             expect(error).toBeInstanceOf(Error);
+    //             expect(error.message).toBe("Employee error");
+    //             expect(events).toBeNull();
+    //             done();
+    //         });
+    //     });
 
-        it("should handle errors from pool.query", (done) => {
-            pool.query.mockImplementationOnce((query, callback) => {
-                callback(new Error("Database error"));
-            });
+    //     it("should handle errors from pool.query", (done) => {
+    //         pool.query.mockImplementationOnce((query, callback) => {
+    //             callback(new Error("Database error"));
+    //         });
 
-            Event.getAllEvents((error, events) => {
-                expect(error).toBeInstanceOf(Error);
-                expect(error.message).toBe("Database error");
-                expect(events).toBeNull();
-                done();
-            });
-        });
-    });
+    //         Event.getAllEvents((error, events) => {
+    //             expect(error).toBeInstanceOf(Error);
+    //             expect(error.message).toBe("Database error");
+    //             expect(events).toBeNull();
+    //             done();
+    //         });
+    //     });
+    // });
 });
 
 describe("getEventById", () => {
@@ -399,122 +399,122 @@ describe("createEvent", () => {
     });
 });
 
-describe("updateEvent", () => {
-    const mockEvent = {
-        id_event: 1,
-        title: "Updated Event",
-        description: "Updated Description",
-        status: 2,
-        is_planned: true,
-        type: "Type 1",
-        id_client: 1,
-        id_address: 1,
-        starting_date: "2023-10-01",
-        starting_hour: "10:00",
-        ending_hour: "12:00",
-        id_employee: 1,
-        work_to_do: "Updated Work",
-    };
+// describe("updateEvent", () => {
+//     const mockEvent = {
+//         id_event: 1,
+//         title: "Updated Event",
+//         description: "Updated Description",
+//         status: 2,
+//         is_planned: true,
+//         type: "Type 1",
+//         id_client: 1,
+//         id_address: 1,
+//         starting_date: "2023-10-01",
+//         starting_hour: "10:00",
+//         ending_hour: "12:00",
+//         id_employee: 1,
+//         work_to_do: "Updated Work",
+//     };
 
-    beforeEach(() => {
-        pool.query = jest.fn((query, values, callback) => {
-            callback(null, { rows: [mockEvent] });
-        });
-    });
+//     beforeEach(() => {
+//         pool.query = jest.fn((query, values, callback) => {
+//             callback(null, { rows: [mockEvent] });
+//         });
+//     });
 
-    it("should update an existing event", (done) => {
-        Event.updateEvent(
-            1,
-            "Updated Event",
-            "Updated Description",
-            2,
-            true,
-            "Type 1",
-            1,
-            1,
-            "2023-10-01",
-            "10:00",
-            "12:00",
-            1,
-            "Updated Work",
-            (error, event) => {
-                expect(error).toBeNull();
-                expect(event).toEqual(
-                    expect.objectContaining({
-                        idEvent: 1,
-                        title: "Updated Event",
-                        description: "Updated Description",
-                        status: 2,
-                        isPlanned: true,
-                        type: "Type 1",
-                        client: expect.objectContaining({
-                            idClient: 1,
-                            category: undefined,
-                            firstname: undefined,
-                            lastname: undefined,
-                            email: undefined,
-                            address: expect.objectContaining({
-                                idAddress: 1,
-                                address: undefined,
-                                city: undefined,
-                                zipcode: undefined,
-                            }),
-                            phoneNumber: undefined,
-                            company: undefined,
-                        }),
-                        address: expect.objectContaining({
-                            idAddress: 1,
-                            address: undefined,
-                            city: undefined,
-                            zipcode: undefined,
-                        }),
-                        startingDate: "2023-10-01",
-                        startingHour: "10:00",
-                        endingHour: "12:00",
-                        employee: expect.objectContaining({
-                            idEmployee: 1,
-                            firstname: undefined,
-                            lastname: undefined,
-                            email: undefined,
-                            phoneNumber: undefined,
-                            position: undefined,
-                        }),
-                        workToDo: "Updated Work",
-                    })
-                );
-                done();
-            }
-        );
-    });
+//     it("should update an existing event", (done) => {
+//         Event.updateEvent(
+//             1,
+//             "Updated Event",
+//             "Updated Description",
+//             2,
+//             true,
+//             "Type 1",
+//             1,
+//             1,
+//             "2023-10-01",
+//             "10:00",
+//             "12:00",
+//             1,
+//             "Updated Work",
+//             (error, event) => {
+//                 expect(error).toBeNull();
+//                 expect(event).toEqual(
+//                     expect.objectContaining({
+//                         idEvent: 1,
+//                         title: "Updated Event",
+//                         description: "Updated Description",
+//                         status: 2,
+//                         isPlanned: true,
+//                         type: "Type 1",
+//                         client: expect.objectContaining({
+//                             idClient: 1,
+//                             category: undefined,
+//                             firstname: undefined,
+//                             lastname: undefined,
+//                             email: undefined,
+//                             address: expect.objectContaining({
+//                                 idAddress: 1,
+//                                 address: undefined,
+//                                 city: undefined,
+//                                 zipcode: undefined,
+//                             }),
+//                             phoneNumber: undefined,
+//                             company: undefined,
+//                         }),
+//                         address: expect.objectContaining({
+//                             idAddress: 1,
+//                             address: undefined,
+//                             city: undefined,
+//                             zipcode: undefined,
+//                         }),
+//                         startingDate: "2023-10-01",
+//                         startingHour: "10:00",
+//                         endingHour: "12:00",
+//                         employee: expect.objectContaining({
+//                             idEmployee: 1,
+//                             firstname: undefined,
+//                             lastname: undefined,
+//                             email: undefined,
+//                             phoneNumber: undefined,
+//                             position: undefined,
+//                         }),
+//                         workToDo: "Updated Work",
+//                     })
+//                 );
+//                 done();
+//             }
+//         );
+//     });
 
-    it("should handle errors from pool.query", (done) => {
-        pool.query.mockImplementationOnce((query, values, callback) => {
-            callback(new Error("Database error"));
-        });
+//     it("should handle errors from pool.query", (done) => {
+//         pool.query.mockImplementationOnce((query, values, callback) => {
+//             callback(new Error("Database error"));
+//         });
 
-        Event.updateEvent(
-            1,
-            "Updated Event",
-            "Updated Description",
-            2,
-            true,
-            "Type 1",
-            1,
-            1,
-            "2023-10-01",
-            "10:00",
-            "12:00",
-            1,
-            "Updated Work",
-            (error, event) => {
-                expect(error).toBeInstanceOf(Error);
-                expect(error.message).toBe("Database error");
-                expect(event).toBeNull();
-                done();
-            }
-        );
-    });
-});
+//         Event.updateEvent(
+//             1,
+//             "Updated Event",
+//             "Updated Description",
+//             2,
+//             true,
+//             "Type 1",
+//             1,
+//             1,
+//             "2023-10-01",
+//             "10:00",
+//             "12:00",
+//             1,
+//             "Updated Work",
+//             (error, event) => {
+//                 expect(error).toBeInstanceOf(Error);
+//                 expect(error.message).toBe("Database error");
+//                 expect(event).toBeNull();
+//                 done();
+//             }
+//         );
+//     });
+// });
 
 describe("deleteEvent", () => {
     const mockEvent = {